skills/whatsapp-voice-chat-integration-open-source/requirements.txt
openai-whisper>=20231117
soundfile>=0.12.1
numpy>=1.21.0
